fix package name mismatches
