+++ android/app/src/main/kotlin/work/racka/reluct/ui/ReluctMainCompose.kt
@@ -1,8 +1,8 @@
 package work.racka.reluct.ui
 
 import androidx.compose.runtime.Composable
-import work.racka.reluct.android.compose.navigation.navhost.AppNavHost
-import work.racka.reluct.android.compose.navigation.util.SettingsCheck
+import work.racka.reluct.android.navigation.navhost.AppNavHost
+import work.racka.reluct.android.navigation.util.SettingsCheck
 import work.racka.reluct.compose.common.theme.ReluctAppTheme
 
 @Composable
