--- android/navigation/src/main/kotlin/work/racka/reluct/android/compose/navigation/navbar/NavBarElevation.kt
@@ -1,12 +1,0 @@
-package work.racka.reluct.android.compose.navigation.navbar
-
-import androidx.compose.ui.unit.dp
-
-internal object NavBarElevation {
-    val Level0 = 0.0.dp
-    val Level1 = 1.0.dp
-    val Level2 = 3.0.dp
-    val Level3 = 6.0.dp
-    val Level4 = 8.0.dp
-    val Level5 = 12.0.dp
-}
