--- android/navigation/src/main/kotlin/work/racka/reluct/android/compose/navigation/toptabs/TabDestination.kt
@@ -1,6 +1,0 @@
-package work.racka.reluct.android.compose.navigation.toptabs
-
-data class TabDestination(
-    val route: String,
-    val name: String
-)
