--- android/navigation/src/main/kotlin/work/racka/reluct/android/compose/navigation/util/SettingsCheck.kt
@@ -1,6 +1,0 @@
-package work.racka.reluct.android.compose.navigation.util
-
-data class SettingsCheck(
-    val isOnBoardingDone: Boolean,
-    val showChangeLog: Boolean
-)
