+++ android/navigation/src/main/kotlin/work/racka/reluct/android/navigation/navbar/NavBarElevation.kt
@@ -1,0 +1,12 @@
+package work.racka.reluct.android.navigation.navbar
+
+import androidx.compose.ui.unit.dp
+
+internal object NavBarElevation {
+    val Level0 = 0.0.dp
+    val Level1 = 1.0.dp
+    val Level2 = 3.0.dp
+    val Level3 = 6.0.dp
+    val Level4 = 8.0.dp
+    val Level5 = 12.0.dp
+}
