+++ android/navigation/src/main/kotlin/work/racka/reluct/android/navigation/navbar/NavbarDestinations.kt
@@ -1,0 +1,53 @@
+package work.racka.reluct.android.navigation.navbar
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Aod
+import androidx.compose.material.icons.outlined.FactCheck
+import androidx.compose.material.icons.outlined.GridView
+import androidx.compose.material.icons.outlined.TaskAlt
+import androidx.compose.material.icons.rounded.Aod
+import androidx.compose.material.icons.rounded.FactCheck
+import androidx.compose.material.icons.rounded.GridView
+import androidx.compose.material.icons.rounded.TaskAlt
+import androidx.compose.ui.graphics.vector.ImageVector
+import work.racka.reluct.common.core_navigation.compose_destinations.dashboard.DashboardOverviewDestination
+import work.racka.reluct.common.core_navigation.compose_destinations.goals.ActiveGoalsDestination
+import work.racka.reluct.common.core_navigation.compose_destinations.screentime.ScreenTimeStatsDestination
+import work.racka.reluct.common.core_navigation.compose_destinations.tasks.PendingTasksDestination
+
+enum class NavbarDestinations(
+    val route: String,
+    val iconActive: ImageVector,
+    val iconInactive: ImageVector,
+    val label: String,
+    val description: String? = null,
+) {
+    Dashboard(
+        route = DashboardOverviewDestination.destination,
+        iconActive = Icons.Rounded.GridView,
+        iconInactive = Icons.Outlined.GridView,
+        label = "Dashboard",
+        description = "Open Dashboard"
+    ),
+    Tasks(
+        route = PendingTasksDestination.destination,
+        iconActive = Icons.Rounded.FactCheck,
+        iconInactive = Icons.Outlined.FactCheck,
+        label = "Tasks",
+        description = "Open Tasks"
+    ),
+    ScreenTime(
+        route = ScreenTimeStatsDestination.destination,
+        iconActive = Icons.Rounded.Aod,
+        iconInactive = Icons.Outlined.Aod,
+        label = "Screen Time",
+        description = "Open Screen Time"
+    ),
+    Goals(
+        route = ActiveGoalsDestination.destination,
+        iconActive = Icons.Rounded.TaskAlt,
+        iconInactive = Icons.Outlined.TaskAlt,
+        label = "Goals",
+        description = "Open Goals"
+    );
+}
