+++ android/navigation/src/main/kotlin/work/racka/reluct/android/navigation/toptabs/TabDestination.kt
@@ -1,0 +1,6 @@
+package work.racka.reluct.android.navigation.toptabs
+
+data class TabDestination(
+    val route: String,
+    val name: String
+)
