+++ android/navigation/src/main/kotlin/work/racka/reluct/android/navigation/toptabs/tasks/TasksTabDestination.kt
@@ -1,0 +1,19 @@
+package work.racka.reluct.android.navigation.toptabs.tasks
+
+import work.racka.reluct.common.core_navigation.compose_destinations.tasks.CompletedTasksDestination
+import work.racka.reluct.common.core_navigation.compose_destinations.tasks.PendingTasksDestination
+import work.racka.reluct.common.core_navigation.compose_destinations.tasks.TasksStatisticsDestination
+
+enum class TasksTabDestination(
+    val route: String
+) {
+    Tasks(
+        route = PendingTasksDestination.route,
+    ),
+    Done(
+        route = CompletedTasksDestination.route,
+    ),
+    Statistics(
+        route = TasksStatisticsDestination.route,
+    );
+}
