+++ android/navigation/src/main/kotlin/work/racka/reluct/android/navigation/util/SettingsCheck.kt
@@ -1,0 +1,6 @@
+package work.racka.reluct.android.navigation.util
+
+data class SettingsCheck(
+    val isOnBoardingDone: Boolean,
+    val showChangeLog: Boolean
+)
